sistergifter/requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-dotenv==1.4.2
django-sendgrid-v5==0.8.0
djangorestframework==3.10.3
future==0.18.2
Pillow==6.2.1
psycopg2==2.8.4
python-http-client==3.2.1
pytz==2019.3
sendgrid==6.1.0
sqlparse==0.3.0